requirements.txt
gradio
requests
numpy
pillow